gradio_frontend/requirements.txt
gradio>=5.49.1
requests>=2.32.5
python-dotenv>=1.0.1